Refactored shape-scooter out of server
